Move unsupported device test to OQD tests
