+++ frontend/test/test_oqd/oqd/test_oqd_device.py
@@ -17,6 +17,7 @@
 import pennylane as qml
 import pytest
 
+from catalyst import qjit
 from catalyst.third_party.oqd import OQDDevice
 
 
@@ -64,6 +65,16 @@ class TestOQDDevice:
         dev = OQDDevice(backend="default", wires=8)
         name, _ = dev.get_c_interface()
         assert name == "oqd"
+
+    def test_unsupported_one_shot_device(self):
+        """Test unsupported device edge case."""
+
+        @qml.qnode(OQDDevice(backend="default", wires=1), shots=10, mcm_method="one-shot")
+        def circuit():
+            return qml.sample()
+
+        with pytest.raises(ValueError, match="'one-shot' is not supported in the chosen device"):
+            qjit(circuit)
 
 
 if __name__ == "__main__":
